requirements.txt
torch==1.13.1
torchvision
torchaudio
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
numpy<2.0.0
matplotlib
scikit-learn
imagebind @ git+https://github.com/BrainVivoDev/ImageBind
pandas
tqdm
open_clip_torch
Pillow
huggingface_hub
lancedb
streamlit
nibabel
dotenv
safetensors